cleanup: remove static symbol facade
