+++ crates/capi/src/dictobject.rs
@@ -1,6 +1,6 @@
 use crate::PyObject;
 use crate::pystate::with_vm;
-use crate::symbols::{exported_object_handle, resolve_object_handle};
+use crate::handles::{exported_object_handle, resolve_object_handle};
 use core::ffi::c_int;
 use rustpython_vm::AsObject;
 use rustpython_vm::builtins::PyDict;
