+++ crates/capi/src/lib.rs
@@ -12,6 +12,7 @@ pub(crate) mod complexobject;
 pub(crate) mod dictobject;
 pub(crate) mod extension_loader;
 pub(crate) mod floatobject;
+pub(crate) mod handles;
 pub(crate) mod import;
 pub(crate) mod listobject;
 pub(crate) mod longobject;
@@ -24,7 +25,6 @@ pub(crate) mod pyerrors;
 pub(crate) mod pylifecycle;
 pub(crate) mod pystate;
 pub(crate) mod refcount;
-pub(crate) mod symbols;
 pub(crate) mod traceback;
 pub(crate) mod tupleobject;
 pub(crate) mod unicodeobject;
